changed difficulty to float64

+++ stratum/client.go
@@ -25,8 +25,8 @@ type StratumClient struct {
 	authorized      bool
 	extraNonce1     string
 	extraNonce2Size uint64
-	prevDifficulty  uint64
-	difficulty      uint64
+	prevDifficulty  float64
+	difficulty      float64
 	remoteAddress   string
 	active          bool
 }
